Replace origin URL with a CrawlJob
Crawls are kicked of with a crawl job description. This includes the URL
the engine starts crawling from and information about which hosts are
considered part of the site that's being crawled. All of this is made
available in the CrawlContext, thus CrawlJob contents can also be used
for filtering crawled docs and links.